Update error handler
