+++ service/http/errors_windows.go
@@ -11,7 +11,7 @@ import (
 
 var errEPIPE = errors.New("WSAECONNABORTED (10053) ->  an established connection was aborted")
 
-// handleWriteError just check if error was caused by aborted connection
+// handleWriteError just check if error was caused by aborted connection on windows
 func handleWriteError(err error) error {
 	if netErr, ok2 := err.(*net.OpError); ok2 {
 		if syscallErr, ok3 := netErr.Err.(*os.SyscallError); ok3 {
