+++ service/http/response.go
@@ -1,22 +1,15 @@
-// +build !windows
-
 package http
 
 import (
-	"errors"
 	"io"
-	"net"
 	"net/http"
-	"os"
 	"strings"
-	"syscall"
 
 	json "github.com/json-iterator/go"
 
 	"github.com/spiral/roadrunner"
 )
 
-var errEPIPE = errors.New("EPIPE(32) -> connection reset by peer")
 
 // Response handles PSR7 response logic.
 type Response struct {
@@ -77,17 +70,6 @@ func (r *Response) Write(w http.ResponseWriter) error {
 	}
 
 	return nil
-}
-
-func handleWriteError(err error) error {
-	if netErr, ok2 := err.(*net.OpError); ok2 {
-		if syscallErr, ok3 := netErr.Err.(*os.SyscallError); ok3 {
-			if syscallErr.Err == syscall.EPIPE {
-				return errEPIPE
-			}
-		}
-	}
-	return err
 }
 
 func handlePushHeaders(h map[string][]string) []string {
